Update docker files
Updates to the Dockerfile to follow the repo-template version.

Avoid needing to copy uv into the prod image by running the prod entrypoint
with just python, not uv run. We're running in the venv created with
uv, using the python version we installed with deadsnakes, so just
python should do the right thing.

Use --frozen rather than --locked to use the uv.lock file as it is.

Also run the tests in the dev entrypoint without uv; dependencies
are installed inthe docker image and we don't need/want it to
re-sync every time